+++ docs/changelog.md
@@ -1,8 +1,10 @@
 # Changelog
 
+## Unreleased
+- add the `ERROR` modlog action
+
 ## v0.10.0
 - add the Gatekeeper cog
-- add the `ERROR` modlog action
 
 ## v0.9.0
 - infraction IDs of members with active infractions (only applies to temprole, tempmute, and mute)